Update README.md
This works by having the user create a login that would be a unique username-password pairing to access their specific file. When they enter the correct information, they can either view or enter new logins to different services or websites. This program is also made for returning and new users.